Separates server.js set up from app.js (good practice)
